Move Java files to related directories
유저 관련 자바 코드들은 User 폴더 안에 넣어주고 게시글 관련 자바 코드들은 Post 폴더 안에 넣어주었습니다.
이렇게 정리하여 한번에 더 알아보기 쉽도록 바꿔주었습니다.

- User, UserService, UserController 자바 파일들을 User 폴더안으로 이동하였습니다.
- Post, PostService, PostController 자바 파일들을 Post 폴더안으로 이동하였습니다.
