+++ src/main/java/codesquad/springcafe/user/User.java
@@ -1,4 +1,4 @@
-package codesquad.springcafe.model;
+package codesquad.springcafe.user;
 
 public class User {
     private Long id;
